Add files via upload
This project studies the patterns and intensity of wildfires and droughts, with a focus on the vulnerability
of California, United States. Observance and analysis of the patterns of disturbance and recovery in vegetation vigour,
water stress and burn damage using the spectral bands and indices that are sensitive to each of the crises
were performed, on the multidimensionally stacked Landsat-8 OLI images over the past decade. After
catching the disturbance vertices which indicate the temporal interval of each environmental disaster
studied